Fixed finalised event page
