+++ packages/frontend/helpers/timeFormatter.tsx
@@ -12,3 +12,38 @@ export function formatTimeRange(
 export function getTZDate(date: Date) {
   return new Date(new Date(date).toISOString().slice(0, 19).replace('Z', ' '));
 }
+
+export function formatTimeBracket(startDate: Date, endDate: Date) {
+  return (
+    startDate.toDateString() +
+    ', ' +
+    (startDate.getHours() < 12
+      ? startDate.getHours() +
+        ':' +
+        startDate.getMinutes() +
+        (startDate.getMinutes() == 0 ? '0' : '') +
+        'am'
+      : (startDate.getHours() == 12
+          ? startDate.getHours()
+          : startDate.getHours() - 12) +
+        ':' +
+        startDate.getMinutes() +
+        (startDate.getMinutes() == 0 ? '0' : '') +
+        'pm') +
+    ' - ' +
+    (endDate.getHours() < 12
+      ? endDate.getHours() +
+        ':' +
+        endDate.getMinutes() +
+        (endDate.getMinutes() == 0 ? '0' : '') +
+        'am'
+      : (endDate.getHours() == 12
+          ? endDate.getHours()
+          : endDate.getHours() - 12) +
+        ':' +
+        endDate.getMinutes() +
+        (endDate.getMinutes() == 0 ? '0' : '') +
+        'pm') +
+    ' NZDT'
+  );
+}
